+++ .env.example
@@ -3,3 +3,7 @@ DISCORD_WEBHOOK_SECRET=
 
 # Sanity webhook secret for authorising revalidate webhook requests
 SANITY_WEBHOOK_SECRET=
+
+# Recaptcha keys
+NEXT_PUBLIC_RECAPTCHA_SITE_KEY=
+RECAPTCHA_SECRET_KEY=
