+++ package.json
@@ -27,6 +27,7 @@
     "next-sanity-image": "^4.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-google-recaptcha-v3": "^1.10.0",
     "react-hook-form": "^7.34.2",
     "react-icons": "^4.4.0",
     "react-spring": "^9.5.2"
