+++ pages/api/contact.ts
@@ -2,8 +2,11 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { ContactFormValues } from "../../lib/contact";
 import { sendContactForm } from "../../lib/contact/discord";
+import { verifyReCaptcha } from "../../lib/recaptcha/verifyReCaptcha";
 
-type RequestData = ContactFormValues;
+interface RequestData extends ContactFormValues {
+  reCaptchaToken: string;
+}
 
 type ApiContactResponse = {
   success: boolean;
@@ -16,30 +19,29 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<ApiContactResponse>
 ) {
-  if (req.method !== "POST") {
-    return res
-      .status(401)
-      .json({ success: false, message: "Method not allowed" });
-  }
-
-  if (!webhookSecret) {
-    return res
-      .status(500)
-      .json({ success: false, message: "Missing webhook secret" });
-  }
+  if (req.method !== "POST") return r(res, 401, "Method not allowed");
+  if (!webhookSecret) return r(res, 500, "Missing webhook secret");
 
   const reqBody: RequestData = req.body;
+
+  const validReCaptcha = await verifyReCaptcha(reqBody.reCaptchaToken);
+  if (!validReCaptcha) return r(res, 401, "Invalid reCaptcha token");
 
   try {
     const result = await sendContactForm(reqBody, webhookSecret);
     if (result) {
-      return res.status(200).json({ success: true, message: "Success" });
+      return r(res, 200, "Success");
     } else {
-      return res
-        .status(500)
-        .json({ success: false, message: "Error sending message" });
+      return r(res, 500, "Error sending message");
     }
   } catch (err) {
-    return res.status(500).send({ success: false, message: "Unknown error" });
+    return r(res, 500, "Unknown error");
   }
 }
+
+// Generate response function
+const r = (
+  res: NextApiResponse<ApiContactResponse>,
+  statusCode: number,
+  message: string
+) => res.status(statusCode).json({ success: statusCode < 300, message });
