+++ styles/globals.css
@@ -59,3 +59,8 @@
 .blue-replace-italic em {
   @apply not-italic text-blue-600 dark:text-blue-400;
 }
+
+.grecaptcha-badge {
+  top: 4rem !important;
+  opacity: 0.7;
+}
